Add comments to QueryBuilderInput
When we open QueryBuilderInput (e.g. Data > Visualise > Series > Who > Dropdown),
we see a console warning like below.

> react-dom.development.js:506 Warning: Using Maps as children is
> unsupported and will likely yield unexpected results.
> Convert it to a sequence/iterable of keyed ReactElements instead.

Type of "selectedOptions" is immutable.List when models.size is 0.
The size of models can be zero when we are opening the dropdown or adding/removing options in the dropdown,
furthermore, it is related to pagination.fetchModels.idsByFilterSelector and state.pagination.

This is just a console warning, not functional bug
Fixing is not simple
So it should be fixed in another branch.